Split nemo loading into compatibility file
